requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.1
cryptography==42.0.5
ecdsa==0.18.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.2
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
mailtrap==2.0.1
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyodbc==5.1.0
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.29
starkbank-ecdsa==2.2.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.29.0
